backend/requirements.txt
fastapi
uvicorn
pyjwt
pymongo
python-dotenv
requests
pytz
